moved reducers and store into different packages
