+++ src/store/configureStore.js
@@ -1,0 +1,13 @@
+import rootReducer from '../reducers'
+import { createStore } from 'redux'
+
+const configureStore = preloadedState => {
+    const store = createStore(
+        rootReducer,
+        preloadedState, 
+    )
+
+    return store
+}
+
+export default configureStore
